pyproject.toml
[project]
name = "workshop-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex",email = "trololo@planet.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (==5.2.4)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
]

[tool.poetry]
packages = [{include = "workshop_backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "django-debug-toolbar (>=6.3.0,<7.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "poetry (>=2.3.3,<3.0.0)",
]
test = [
    "coverage (>=7.13.5,<8.0.0)",
    "pytest (>=9.0.3,<10.0.0)"
]
build = [
    "poetry-plugin-export (>=1.10.0,<2.0.0)"
]
